+++ CHANGELOG.rst
@@ -4,10 +4,12 @@ Toolium Changelog
 v1.2.3
 ------
 
-*In development*
+*Release date: 2017-03-10*
 
 - Save *geckodriver.log* file in output folder
 - Fix MagickEngine name error when using an old version of needle
+- Add *wait_until_clickable* method to Utils and PageElement classes to search for an element and wait until it is
+  clickable
 
 v1.2.2
 ------